Extracted loaders to seperate modules.
