+++ packages/ckeditor5-dev-utils/lib/loaders/getformattedtextloader.js
@@ -1,0 +1,14 @@
+/**
+ * @license Copyright (c) 2003-2024, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+/**
+ * @returns {Object}
+ */
+export default function getFormattedTextLoader() {
+	return {
+		test: /\.(txt|html|rtf)$/,
+		use: [ 'raw-loader' ]
+	};
+}
