+++ packages/ckeditor5-dev-utils/tests/loaders/getcoverageloader.js
@@ -1,0 +1,82 @@
+/**
+ * @license Copyright (c) 2003-2024, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+import { describe, expect, it, vi } from 'vitest';
+
+
+describe( 'getCoverageLoader()', () => {
+	it( 'should be a function', () => {
+		expect( loaders.getCoverageLoader ).to.be.a( 'function' );
+	} );
+
+	it( 'should return a definition containing a loader for measuring the coverage', () => {
+		const coverageLoader = loaders.getCoverageLoader( {
+			files: []
+		} );
+
+		expect( coverageLoader ).to.be.an( 'object' );
+		expect( '/path/to/javascript.js' ).to.match( coverageLoader.test );
+		expect( '/path/to/typescript.ts' ).to.match( coverageLoader.test );
+
+		expect( coverageLoader.include ).to.be.an( 'array' );
+		expect( coverageLoader.include ).to.lengthOf( 0 );
+		expect( coverageLoader.exclude ).to.be.an( 'array' );
+		expect( coverageLoader.exclude ).to.lengthOf( 1 );
+
+		expect( coverageLoader.use ).to.be.an( 'array' );
+		expect( coverageLoader.use ).to.lengthOf( 1 );
+
+		const babelLoader = coverageLoader.use[ 0 ];
+
+		expect( babelLoader.loader ).to.equal( 'babel-loader' );
+	} );
+
+	it( 'should return a definition containing a loader for measuring the coverage (include glob check)', () => {
+		const coverageLoader = loaders.getCoverageLoader( {
+			files: [
+				// -f utils
+				[ 'node_modules/ckeditor5-utils/tests/**/*.js' ]
+			]
+		} );
+
+		expect( coverageLoader ).to.be.an( 'object' );
+		expect( coverageLoader ).to.have.property( 'include' );
+		expect( coverageLoader.include ).to.be.an( 'array' );
+		expect( coverageLoader.include ).to.deep.equal( [
+			new RegExp( [ 'ckeditor5-utils', 'src', '' ].join( escapedPathSep ) )
+		] );
+	} );
+
+	it( 'should return a definition containing a loader for measuring the coverage (exclude glob check)', () => {
+		const coverageLoader = loaders.getCoverageLoader( {
+			files: [
+				// -f !utils
+				[ 'node_modules/ckeditor5-!(utils)/tests/**/*.js' ]
+			]
+		} );
+
+		expect( coverageLoader ).to.be.an( 'object' );
+		expect( coverageLoader ).to.have.property( 'include' );
+		expect( coverageLoader.include ).to.be.an( 'array' );
+		expect( coverageLoader.include ).to.deep.equal( [
+			new RegExp( [ 'ckeditor5-!(utils)', 'src', '' ].join( escapedPathSep ) )
+		] );
+	} );
+
+	it( 'should return a definition containing a loader for measuring the coverage (for root named ckeditor5-*)', () => {
+		const coverageLoader = loaders.getCoverageLoader( {
+			files: [
+				[ '/ckeditor5-collab/packages/ckeditor5-alignment/tests/**/*.{js,ts}' ]
+			]
+		} );
+
+		expect( coverageLoader ).to.be.an( 'object' );
+		expect( coverageLoader ).to.have.property( 'include' );
+		expect( coverageLoader.include ).to.be.an( 'array' );
+		expect( coverageLoader.include ).to.deep.equal( [
+			new RegExp( [ 'ckeditor5-alignment', 'src', '' ].join( escapedPathSep ) )
+		] );
+	} );
+} );
